+++ Source/Core/TTY.h
@@ -15,7 +15,7 @@ namespace TTY
 void Write(const std::string& string);
 void Write(const char c);
 void Scroll(const u8 lines, const u8 color);
-void MoveCursor(const u32 x, const u32 y);
+void MoveCursor(const u8 x, const u8 y);
 u8 GetCursorRow();
 void SetCursorRow(u8 row);
 u8 GetCursorColumn();
